Update helpers.py
There was added landmark of inner_mouth, so the number of landmarks became 8. There was an issue in function visualize_facial_landmarks when colors were set to None. I added a new color to inner  inner_mouth - dark green in BGR (31, 47, 0) at line 65. It is only propose, the color can be any relevant.